refactor src to move into `show_dialog`
