+++ setup.cfg
@@ -1,7 +1,7 @@
 # Configuration for tools that don't support `pyproject.toml`.
 
 [flake8]
-extend-exclude = src/ui/forms,.venv*,venv*,.git,.github,assets,dist,docs,site,__pycache__
+extend-exclude = src/show_dialog/ui/forms,.venv*,venv*,.git,.github,assets,dist,docs,site,__pycache__
 max-line-length = 100
 # Errors being ignored:
 # E203 is not PEP8 compliant and clashes with black
@@ -35,7 +35,7 @@ classifiers =
 
 [options]
 package_dir =
-    = src
+    = src/show_dialog
 packages = find:
 python_requires = >=3.10
 include_package_data = True
@@ -44,7 +44,7 @@ include_package_data = True
 show_dialog = assets/stylesheets/*.css, assets/app.yaml
 
 [options.packages.find]
-where = src
+where = src/show_dialog
 exclude =
     tests
     tests.*
