+++ src/show_dialog/config.py
