+++ src/show_dialog/ui/forms/resources_rc.py
